refactor(#75): 채널 보안 SRP 구조 정리
