+++ src/app/channel/security/page.tsx
@@ -1,6 +1,7 @@
 // 채널 보안 설정 페이지를 렌더링합니다.
-import { getChannelSecuritySnapshot } from "@/actions/channel/security";
 import ChannelSecurityPageClient from "@/components/channel/security/channel-security-page-client";
+
+import { getChannelSecuritySnapshot } from "./data";
 
 export default async function ChannelSecurityPage() {
   const result = await getChannelSecuritySnapshot();
