+++ src/app/live/[creatorId]/alerts/donation/[overlayKey]/page.tsx
@@ -1,7 +1,8 @@
 // OBS 후원 알림 URL의 보안 key 경로를 받는 화면을 렌더링합니다.
-import { getLiveDonationAlertOverlaySnapshot } from "@/actions/live/overlay";
 import { LiveDonationAlertOverlay } from "@/components/live/overlay/live-donation-alert-overlay";
 import { notFound } from "next/navigation";
+
+import { getLiveDonationAlertOverlaySnapshot } from "../../../_data/live-overlay-data";
 
 export const metadata = {
   title: "후원 알림",
